+++ src/Shared/DC/ZRDB/tests/test_sqlgroup.py
@@ -197,5 +197,6 @@ update actors
 
 def test_suite():
     suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(SQLGroupTests))
+    suite.addTest(
+        unittest.defaultTestLoader.loadTestsFromTestCase(SQLGroupTests))
     return suite